+++ EntraIdDSC/public/Remove-EntraIdGroup.ps1
@@ -17,41 +17,45 @@
     None
 #>
 function Remove-EntraIdGroup {
-    [CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = 'High')]
+    [CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = 'High', DefaultParameterSetName = 'ById')]
     param (
-        [Parameter(Mandatory = $false, ValueFromPipelineByPropertyName = $true)]
-        [ValidateNotNullOrEmpty()]
-        [string]$GroupId,
+        [Parameter(ParameterSetName = 'ById', ValueFromPipelineByPropertyName = $true)]
+        [string]$Id,
 
-        [Parameter(Mandatory = $false, ValueFromPipelineByPropertyName = $true)]
-        [ValidateNotNullOrEmpty()]
-        [string]$GroupName
+        [Parameter(ParameterSetName = 'ByName', ValueFromPipelineByPropertyName = $true)]
+        [string]$DisplayName
     )
 
     process {
-        try {
-            if ($GroupId) {
-                if ($PSCmdlet.ShouldProcess($GroupId, "Remove Entra ID group by Id")) {
-                    Remove-MgGroup -GroupId $GroupId -ErrorAction Stop
-                    Write-Verbose "Group with Id '$GroupId' removed."
+        switch ($PSCmdlet.ParameterSetName) {
+            'ById' {
+                if (-not $Id) {
+                    Throw "Id cannot be empty."
+                }
+                if ($PSCmdlet.ShouldProcess($Id, "Remove Entra ID group by Id")) {
+                    Remove-MgGroup -GroupId $Id -ErrorAction Stop
+                    Write-Verbose "Group with Id '$Id' removed."
                 }
             }
-            elseif ($GroupName) {
-                $group = Get-MgGroup -Filter "displayName eq '$GroupName'" -ErrorAction Stop
+            'ByName' {
+                if (-not $DisplayName) {
+                    Throw "DisplayName cannot be empty."
+                }
+                $group = Get-MgGroup -Filter "displayName eq '$DisplayName'" -ErrorAction Stop
                 if ($group) {
-                    if ($PSCmdlet.ShouldProcess($GroupName, "Remove Entra ID group by Name")) {
+                    if ($PSCmdlet.ShouldProcess($DisplayName, "Remove Entra ID group by Name")) {
                         Remove-MgGroup -GroupId $group.Id -ErrorAction Stop
-                        Write-Verbose "Group with Name '$GroupName' removed."
+                        Write-Verbose "Group with Name '$DisplayName' removed."
                     }
-                } else {
-                    Write-Error "Group with Name '$GroupName' not found."
+                }
+                else {
+                    Throw "Group with Name '$DisplayName' not found."
                 }
             }
-            else {
-                Write-Error "You must specify either GroupId or GroupName."
+            default {
+                Throw "You must specify either Id or DisplayName."
             }
-        } catch {
-            Write-Error $_
         }
+
     }
 }